Minor refactoring of Dictionary extensions
Note that the 'toArray(_:)' method seemingly currently duplicates the
functionality of the 'map(:_)' method of Dictionary, available via a
default implementation from 'Collection'. This fact is quite apparent
after the re-factoring of this commit, and possibly prompts a future
deprecation of the 'toArray(_:)' method, unless it's deemed valuable as
a utility method for pure semantic reasons.

Swiftlint @ DictionaryExtensions OK (unchanged)